// front/src/shared/api/generated/model/orderInDB.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInDB } from './orderItemInDB';

export interface OrderInDB {
  fullName: string;
  email: string;
  address: string;
  city: string;
  postalCode: string;
  phone: string;
  totalAmount: string;
  orderItems?: OrderItemInDB[];
  id: string;
  userId: string;
  status: string;
  createdAt: string;
}